fix: move types to libs folder gf-53
