+++ packages/shared/src/libs/helpers/change-case/change-case.helper.ts
@@ -1,5 +1,5 @@
 import { caseTypeToFunction } from "./libs/maps/maps.js";
-import { type CaseType } from "./types/types.js";
+import { type CaseType } from "./libs/types/types.js";
 
 const changeCase = (input: string, caseType: CaseType): string => {
 	const caseFunction = caseTypeToFunction[caseType];
