+++ packages/shared/src/libs/helpers/change-case/libs/maps/case-type-to-function.map.ts
@@ -1,6 +1,6 @@
 import { camelCase, snakeCase } from "change-case";
 
-import { type CaseType } from "../../types/types.js";
+import { type CaseType } from "../types/types.js";
 
 const caseTypeToFunction: Record<CaseType, (input: string) => string> = {
 	camelCase,
