+++ packages/shared/src/libs/helpers/change-case/libs/types/case-type.type.ts
